fix: 推理bf16作用域/词表一致性/int16溢出等找bug修复+回归测试
- generate.py: bf16 由全局 set_autocast 改为 AMP_CTX 上下文管理器，仅包裹前向/生成，
  避免泄漏到采样算子且永不关闭（与 train.py 一致）；DML 后端退化为 nullcontext
- data_utils.py: TextDataset 预编码张量 int16->int32，修复 vocab_size>32767 时 token id 回绕
- train.py/checkpoint.py: 新增 model.vocab_size >= 分词器词表大小 断言，防止 token id 越界
- generate.py: _temp_applied 增加 _ngram_fusion_active 判断，覆盖运行时关闭融合的边界
- 证伪两个误报（训练期/流畅度 double-log_softmax）：fused=log_softmax(z)+prior 的
  -LSE(z) 为逐行常数，softmax 平移不变使 CE(fused)==CE(z+prior)；且 F.log_softmax(fused)
  恰好抵消 prior 偏移得正确组合分布，故 _fluency_batch 的 log_softmax 必须保留
- 新增回归测试 test_ngram_fusion_fluency_requires_log_softmax / test_fluency_batch_applies_log_softmax
- pytest 138 passed / 2 skipped

+++ models/data_utils.py
@@ -36,7 +36,8 @@ class TextDataset(Dataset):
         # “GPU 不响应更多命令 / device reset”导致 backward 崩溃。
         L = max_seq_length + 1
         pad = self.vocab.pad_idx
-        arr = np.zeros((len(texts), L), dtype=np.int16)
+        # int32 以容纳 vocab_size > 32767 的词表（int16 上限 32767，超界会回绕静默损坏 token id）。
+        arr = np.zeros((len(texts), L), dtype=np.int32)
         for i, text in enumerate(texts):
             tokens = self.vocab.encode(text)
             n = len(tokens)
@@ -45,7 +46,7 @@ class TextDataset(Dataset):
             elif n < L:
                 tokens = tokens + [pad] * (L - n)
             arr[i] = tokens
-        self.tokens = torch.from_numpy(arr)  # (N, L) int16，取用时转 long
+        self.tokens = torch.from_numpy(arr)  # (N, L) int32，取用时转 long
         self.texts = None  # 释放原始字符串，回收内存
         
     def _preprocess_texts(self, texts: List[str]) -> List[str]:
